chore: merge alive page into status
